+++ src/IotHub/IotHub/ChangeLog.md
@@ -19,6 +19,9 @@
 -->
 ## Upcoming Release
 
+## Version 2.7.1
+* Fixed an issue of SAS token.
+
 ## Version 2.7.0
 * Allowed tags in IoT Hub create cmdlet.